+++ src/engineering/__init__.py
@@ -1,0 +1,4 @@
+"""Engineering authority bundle namespace.
+
+Author: © 2026 Afshin Arjhangmehr
+"""